random: one CRuby-exact Mersenne Twister for the global PRNG
The seeded global PRNG behind Kernel#srand / #rand / Random.rand /
Random.bytes was a rand_mt::Mt seeded with the seed's low 32 bits and
drawn via rand's generic f64 distribution, so none of its streams
matched CRuby even though Random *instances* already carried a
bit-exact MT19937. Move that reference implementation (init_genrand /
init_by_array seeding from every word of the seed, genrand_real,
limited_big_rand rejection sampling, rand_int, rb_rand_bytes) into
globals::prng, make the global Prng use it, and re-export the helpers
to builtins::random so both RNGs share one machinery.

Kernel#rand now draws exactly like CRuby: rb_random_ulong_limited for
integer maxima and integer ranges (no more float-scaling bias), the
53-bit genrand_real for floats, int_pair_to_real_inclusive for
*inclusive* float ranges, and the multi-word rand_int path for Bignum
maxima (rand(2**80) raised RangeError before). Random.bytes uses
rb_rand_bytes' little-endian word protocol.

With the draws unified, `srand(123); ary.shuffle` produces CRuby's
exact sequence (the shuffle side lands in the following commit).
